# backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Junkio Expense Tracker API
  version: 1.0.0
  description: >-
    **Chào mừng đến với thư viện API của Junkio Expense Tracker!**  


    Dưới đây là các tài liệu hướng dẫn sử dụng API chi tiết dành cho Lập trình
    viên, Tester và Admin hệ thống. Thiết kế tuân theo chuẩn RESTful.


    ---

    ###  Hướng dẫn dành cho Người Dùng Mới (Getting Started)


    Nếu bạn là người mới sử dụng API này, hãy làm theo các bước sau để xác thực:


    1. **Đăng nhập (Login):** Go to the `/api/auth/login` route và nhập Email &
    Password.

    2. **Lấy Token:** Copy chuỗi `token` trả về từ kết quả JSON (trong object
    `data`).

    3. **Cấp quyền (Authorize):** Kéo lên trên cùng của trang web này, bấm vào
    nút **Authorize** màu xanh lá cây hoặc click vào biểu tượng 🔒 ở bất kỳ API
    nào, dán chuỗi token vừa copy vào ô **Value** và nhấn **Authorize**.

    4. **Bắt đầu gọi API:** Nhấn `Try it out` ở các endpoint có yêu cầu xác thực
    để thực hiện các yêu cầu (Requests). Server sẽ nhận diện được phiên làm việc
    của bạn.


    > ** Mẹo:** Swagger UI đã được cấu hình lưu lại Token kể cả khi bạn tải lại
    trang (Persist Authorization).


    ---


    ###  Mã Lỗi (Error Codes)

    Junkio API sử dụng hệ thống `Error Codes` chuẩn và thông báo lỗi đa ngôn ngữ
    (i18n). Thay vì đọc chuỗi ký tự thô, Frontend sẽ nhận được các mã như
    `WALLET_NOT_FOUND`, `INSUFFICIENT_BALANCE` và tự thông dịch thành văn bản.
...
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /api/transactions:
    get:
      summary: Lấy danh sách giao dịch có phân trang và bộ lọc
      description: >
        Trả về danh sách giao dịch của người dùng hiện tại, hỗ trợ nhiều bộ lọc:

        - **Context**: Xem giao dịch cá nhân (`personal`) hoặc gia đình
        (`family`)

        - **Loại giao dịch**: INCOME, EXPENSE, TRANSFER_IN, TRANSFER_OUT

        - **Khoảng thời gian**: Lọc theo ngày bắt đầu / kết thúc

        - **Ví cụ thể**: Lọc theo wallet_id

        - **Danh mục**: Lọc theo category_id

        - **Tìm kiếm**: Tìm theo mô tả giao dịch


        Kết quả có phân trang, mặc định 10 bản ghi mỗi trang.
      tags:
        - Transactions
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: context
          schema:
            type: string
            enum:
              - personal
              - family
          description: Chọn ngữ cảnh dữ liệu cần xem (cá nhân hoặc gia đình)
        - in: query
          name: family_id
          schema:
            type: string
            format: uuid
          description: ID gia đình (bắt buộc khi context = family)
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Số trang hiện tại (bắt đầu từ 1)
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Số giao dịch mỗi trang
        - in: query
          name: type
          schema:
            type: string
            enum:
              - INCOME
              - EXPENSE
              - TRANSFER_IN
              - TRANSFER_OUT
          description: Lọc theo loại giao dịch
        - in: query
          name: startDate
          schema:
            type: string
            format: date
          description: Ngày bắt đầu lọc (YYYY-MM-DD)
        - in: query
          name: endDate
          schema:
            type: string
            format: date
          description: Ngày kết thúc lọc (YYYY-MM-DD)
        - in: query
          name: wallet_id
          schema:
            type: string
            format: uuid
          description: Lọc theo ID ví cụ thể
        - in: query
          name: category_id
          schema:
            type: string
            format: uuid
          description: Lọc theo danh mục thu/chi
        - in: query
          name: search
          schema:
            type: string
          description: Tìm kiếm theo từ khóa trong mô tả giao dịch
      responses:
        '200':
          description: Danh sách giao dịch thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      transactions:
                        type: array
                        items:
                          type: object
                      totalItems:
                        type: integer
                      totalPages:
                        type: integer
                      currentPage:
                        type: integer
              example:
                status: success
                message: Lấy danh sách giao dịch thành công
                data:
                  transactions:
                    - id: t1a2b3c4-5678-9abc-def0-123456789001
                      amount: 150000
                      type: EXPENSE
                      description: Cà phê buổi sáng
                      date: '2026-03-30'
                      wallet_id: w1a2b3c4-...
                      category_id: c1a2b3c4-...
                    - id: t1a2b3c4-5678-9abc-def0-123456789002
                      amount: 25000000
                      type: INCOME
                      description: Lương tháng 3
                      date: '2026-03-28'
                      wallet_id: w1a2b3c4-...
                      category_id: null
                  totalItems: 342
                  totalPages: 35
                  currentPage: 1
    post:
      summary: Tạo giao dịch mới (thu nhập hoặc chi tiêu)
      description: >
        Ghi nhận một giao dịch thu nhập hoặc chi tiêu mới vào ví.


        **Yêu cầu:**

        - Phải có ít nhất một ví hợp lệ trước khi tạo giao dịch

        - `wallet_id` phải là ví mà bạn sở hữu hoặc có quyền truy cập (ví gia
        đình)

        - Khi tạo giao dịch EXPENSE, số dư ví phải đủ


        **Luồng xử lý:**

        - INCOME: Tự động cộng `amount` vào số dư ví

        - EXPENSE: Tự động trừ `amount` khỏi số dư ví
      tags:
        - Transactions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - wallet_id
                - amount
                - type
                - date
              properties:
                wallet_id:
                  type: string
                  format: uuid
                  description: ID ví thực hiện giao dịch
                amount:
                  type: number
                  minimum: 0.01
                  example: 150000
                  description: Số tiền giao dịch (phải lớn hơn 0)
                type:
                  type: string
                  enum:
                    - INCOME
                    - EXPENSE
                  example: EXPENSE
                  description: Loại giao dịch
                description:
                  type: string
                  example: Cà phê buổi sáng
                  description: Mô tả giao dịch (tùy chọn)
                category_id:
                  type: string
                  format: uuid
                  nullable: true
                  description: ID danh mục (tùy chọn)
                date:
                  type: string
                  format: date
                  example: '2026-03-30'
                  description: Ngày giao dịch (YYYY-MM-DD)
            examples:
              chi_tieu:
                summary: Ghi nhận chi tiêu
                value:
                  wallet_id: w1a2b3c4-5678-9abc-def0-123456789001
                  amount: 150000
                  type: EXPENSE
                  description: Cà phê buổi sáng
                  date: '2026-03-30'
              thu_nhap:
                summary: Ghi nhận thu nhập
                value:
                  wallet_id: w1a2b3c4-5678-9abc-def0-123456789001
                  amount: 25000000
                  type: INCOME
                  description: Lương tháng 3/2026
                  date: '2026-03-28'
      responses:
        '201':
          description: Tạo giao dịch thành công
          content:
            application/json:
              example:
                status: success
                message: Tạo giao dịch thành công
                data:
                  id: t1a2b3c4-new-transaction-id
                  amount: 150000
                  type: EXPENSE
                  description: Cà phê buổi sáng
                  wallet_balance: 9850000
        '400':
          description: Chưa có ví hoặc số dư không đủ (INSUFFICIENT_BALANCE)
        '422':
          description: Dữ liệu body không hợp lệ
  /api/transactions/transfer:
    post:
      summary: Chuyển tiền giữa hai ví
      description: >
        Thực hiện chuyển tiền từ ví nguồn sang ví đích.

        Hệ thống tự động tạo 2 giao dịch liên kết:

        - **TRANSFER_OUT** trên ví nguồn (trừ tiền)

        - **TRANSFER_IN** trên ví đích (cộng tiền)


        Cả hai ví phải thuộc quyền quản lý của user (ví cá nhân hoặc ví gia
        đình).

        Hai giao dịch được liên kết bằng `transfer_group_id` để dễ truy vết.
      tags:
        - Transactions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - from_wallet_id
                - to_wallet_id
                - amount
              properties:
                from_wallet_id:
                  type: string
                  format: uuid
                  description: ID ví nguồn (trừ tiền)
                to_wallet_id:
                  type: string
                  format: uuid
                  description: ID ví đích (cộng tiền)
                amount:
                  type: number
                  minimum: 0.01
                  example: 5000000
                  description: Số tiền chuyển (phải nhỏ hơn hoặc bằng số dư ví nguồn)
                description:
                  type: string
                  example: Chuyển tiền tiết kiệm
                  description: Ghi chú chuyển khoản
                date:
                  type: string
                  format: date
                  description: Ngày chuyển khoản
            example:
              from_wallet_id: w-source-1234-...
              to_wallet_id: w-dest-5678-...
              amount: 5000000
              description: Chuyển sang ví tiết kiệm
      responses:
        '201':
          description: Chuyển tiền thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      transfer_group_id:
                        type: string
                        format: uuid
                      transfer_out_id:
                        type: string
                        format: uuid
                      transfer_in_id:
                        type: string
                        format: uuid
                      from_wallet_balance:
                        type: number
                      to_wallet_balance:
                        type: number
              example:
                status: success
                message: Chuyển tiền thành công
                data:
                  transfer_group_id: tg-a1b2c3d4-...
                  transfer_out_id: t-out-5678-...
                  transfer_in_id: t-in-9012-...
                  from_wallet_balance: 5000000
                  to_wallet_balance: 15000000
        '400':
          description: Ví không hợp lệ hoặc số dư không đủ (INSUFFICIENT_BALANCE)
  /api/transactions/import:
    post:
      summary: Nhập hàng loạt giao dịch (Bulk Import)
      description: |
        Cho phép nhập nhiều giao dịch cùng lúc trong một request.
        Hữu ích khi:
        - Chuyển dữ liệu từ ứng dụng khác sang Junkio
        - Nhập dữ liệu từ file Excel/CSV đã parse
        - Tự động hóa việc ghi nhận giao dịch

        **Lưu ý:** Mỗi giao dịch trong mảng sẽ được xử lý độc lập.
        Nếu một giao dịch lỗi, các giao dịch khác vẫn được tạo thành công.
      tags:
        - Transactions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - transactions
              properties:
                transactions:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required:
                      - wallet_id
                      - amount
                      - type
                      - date
                    properties:
                      wallet_id:
                        type: string
                        format: uuid
                      amount:
                        type: number
                      type:
                        type: string
                        enum:
                          - INCOME
                          - EXPENSE
                      description:
                        type: string
                      category_id:
                        type: string
                        format: uuid
                        nullable: true
                      date:
                        type: string
                        format: date
            example:
              transactions:
                - wallet_id: w1a2b3c4-...
                  amount: 50000
                  type: EXPENSE
                  description: Grab đi làm
                  date: '2026-03-28'
                - wallet_id: w1a2b3c4-...
                  amount: 200000
                  type: EXPENSE
                  description: Mua sách
                  date: '2026-03-29'
      responses:
        '200':
          description: Nhập dữ liệu thành công
          content:
            application/json:
              example:
                status: success
                message: Nhập 2 giao dịch thành công
        '400':
          description: Body request rỗng hoặc bạn chưa có ví hợp lệ
...